refact(TrialExtensionForm): extract useSignAndPay into its own module
Co-authored: @robinandeer
