+++ apps/store/src/features/carDealership/TrialExtensionForm.tsx
@@ -1,22 +1,13 @@
-import { datadogLogs } from '@datadog/browser-logs'
 import { datadogRum } from '@datadog/browser-rum'
-import { useRouter } from 'next/router'
 import { useState } from 'react'
 import { Space, Button, RestartIcon, Text, BankIdIcon } from 'ui'
 import { InfoCard } from '@/components/InfoCard/InfoCard'
 import { ProductItemContainer } from '@/components/ProductItem/ProductItemContainer'
 import { SpaceFlex } from '@/components/SpaceFlex/SpaceFlex'
-import { Cart, ShopSessionAuthenticationStatus } from '@/services/apollo/generated'
-import { useCurrentMemberLazyQuery } from '@/services/apollo/generated'
-import { useAppErrorHandleContext } from '@/services/appErrors/AppErrorContext'
-import { useBankIdContext } from '@/services/bankId/BankIdContext'
-import { PageLink } from '@/utils/PageLink'
-import { useAddToCart } from '@/utils/useAddToCart'
 import { ActionButtonsCar } from './ActionButtonsCar'
 import { type CarTrialData } from './carDealershipFixtures'
 import { ProductItemContractContainerCar } from './ProductItemContractContainer'
-
-const carDealershipLogger = datadogLogs.createLogger('car-dealership')
+import { useSignAndPay } from './useSignAndPay'
 
 const SIGN_AND_PAY_BUTTON = 'Sign and pay'
 const SIGN_BUTTON = 'Sign insurance'
@@ -121,79 +112,4 @@ export const TrialExtensionForm = (props: Props) => {
       </Space>
     </Space>
   )
-}
-
-type Params = {
-  shopSessionId: string
-  ssn: string
-  authenticationStatus: ShopSessionAuthenticationStatus
-  cartEntries: Cart['entries']
-  requirePaymentConnection: boolean
-}
-
-const useSignAndPay = (params: Params) => {
-  const { startCheckoutSign } = useBankIdContext()
-  const router = useRouter()
-  const { showError } = useAppErrorHandleContext()
-
-  const [getCurrentMember] = useCurrentMemberLazyQuery({
-    onCompleted(data) {
-      if (data.currentMember.hasActivePaymentConnection) {
-        carDealershipLogger.info('Member has active payment connection', {
-          promptedPayment: params.requirePaymentConnection,
-        })
-        router.push(PageLink.confirmation({ shopSessionId: params.shopSessionId }))
-      } else {
-        carDealershipLogger.info('Member does not have active payment connection', {
-          promptedPayment: params.requirePaymentConnection,
-        })
-        router.push(PageLink.checkoutPaymentTrustly({ shopSessionId: params.shopSessionId }))
-      }
-    },
-  })
-
-  const performSign = () => {
-    startCheckoutSign({
-      shopSessionId: params.shopSessionId,
-      ssn: params.ssn,
-      customerAuthenticationStatus: params.authenticationStatus,
-
-      onSuccess() {
-        console.info('Successfully signed shop session')
-        getCurrentMember()
-      },
-    })
-  }
-
-  const [addToCart, loadingAddToCart] = useAddToCart({
-    shopSessionId: params.shopSessionId,
-    onSuccess(productOfferId) {
-      console.info('Successfully added product to cart', productOfferId)
-      performSign()
-    },
-  })
-
-  const addAndOrSign = (offerId: string) => {
-    if (params.cartEntries.length > 1) {
-      showError(
-        new Error(
-          `Cart has unexpected items in it. cartOfferIds=${params.cartEntries.map(
-            ({ id }) => id,
-          )}. Offer to be added: ${offerId}`,
-        ),
-      )
-      return
-    }
-
-    const alreadyAdded = params.cartEntries.some((entry) => entry.id === offerId)
-    if (alreadyAdded) {
-      carDealershipLogger.info('Offer already added to cart')
-      performSign()
-    } else {
-      carDealershipLogger.info(`Adding offer to cart: ${offerId}`)
-      addToCart(offerId)
-    }
-  }
-
-  return { signAndPay: addAndOrSign, loading: loadingAddToCart } as const
 }
